+++ android/src/main/java/flutter/overlay/window/flutter_overlay_window/WindowSetup.java
@@ -15,25 +15,11 @@ public abstract class WindowSetup {
     static int flag = WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE;
     static int gravity = Gravity.CENTER;
     static BasicMessageChannel<Object> messenger = null;
-    static String overlayTitle = "Overlay is activated";
-    static String overlayContent = "Tap to edit settings or disable";
-    static String notificationIcon = null;
     static String positionGravity = "none";
-    static int notificationVisibility = NotificationCompat.VISIBILITY_PRIVATE;
     static boolean enableDrag = false;
 
 
-    static void setNotificationVisibility(String name) {
-        if (name.equalsIgnoreCase("visibilityPublic")) {
-            notificationVisibility = NotificationCompat.VISIBILITY_PUBLIC;
-        }
-        if (name.equalsIgnoreCase("visibilitySecret")) {
-            notificationVisibility = NotificationCompat.VISIBILITY_SECRET;
-        }
-        if (name.equalsIgnoreCase("visibilityPrivate")) {
-            notificationVisibility = NotificationCompat.VISIBILITY_PRIVATE;
-        }
-    }
+    // Notification settings removed (no foreground service)
 
     static void setFlag(String name) {
         if (name.equalsIgnoreCase("flagNotFocusable") || name.equalsIgnoreCase("defaultFlag")) {