+++ lib/src/overlay_config.dart
@@ -56,16 +56,7 @@ enum OverlayFlag {
 }
 
 /// The level of detail displayed in notifications on the lock screen.
-enum NotificationVisibility {
-  /// Show this notification in its entirety on all lockscreens.
-  visibilityPublic,
-
-  /// Do not reveal any part of this notification on a secure lockscreen.
-  visibilitySecret,
-
-  /// Show this notification on all lockscreens, but conceal sensitive or private information on secure lockscreens.
-  visibilityPrivate
-}
+// Notification visibility is no longer used since overlay doesn't run as a foreground service.
 
 class WindowSize {
   WindowSize._();